KS-13-1 -- Contacts : nullable field

+++ src/main/kotlin/lesson_13/lesson_13-task_1.kt
@@ -4,6 +4,6 @@ package lesson_13
 
 class Contact1(
     val name: String,
-    val phoneNumber: Int,
+    val phoneNumber: Long,
     val company: String?,
 )
